+++ apps/storybook/.storybook/theme.js
@@ -1,13 +1,16 @@
 import { create } from "@storybook/theming";
 import logo from "./kvib_logo.svg";
+import "@fontsource-variable/mulish";
+
+const isLocalhost = window.location.hostname === "localhost";
 
 export default create({
   base: "light",
   // Typography
-  fontBase: '"Mulish", "Open Sans", sans-serif',
+  fontBase: '"Mulish Variable", "Open Sans", sans-serif',
   fontCode: "monospace",
   brandTitle: "KVIB",
-  brandUrl: "/",
+  brandUrl: isLocalhost ? "/" : "/kvib",
   brandImage: logo,
   brandTarget: "_self",