+++ packages/react/src/footer/Footer.tsx
@@ -7,6 +7,9 @@ export type FooterProps = {
   excludeContactInfo?: boolean;
   excludeHelp?: boolean;
   excludeNews?: boolean;
+  /**The contact info email-adress
+   * @default post@kartverket.no*/
+  contactInfoEmailAddress?: string;
 };
 
 const FooterToggleableFlex = (props: FlexProps) => {
@@ -26,6 +29,7 @@ export const Footer = ({
   excludeHelp,
   excludeNews,
   excludeSocialMedia,
+  contactInfoEmailAddress = "post@kartverket.no",
 }: FooterProps) => {
   const onlyOneIncluded =
     (!excludeContactInfo || !excludeOpeningHours ? 1 : 0) + (!excludeHelp ? 1 : 0) + (!excludeNews ? 1 : 0) <= 1;
@@ -50,9 +54,13 @@ export const Footer = ({
                 </Text>
                 <Text>
                   E-post:{" "}
-                  <Link fontWeight="bold" aria-label="Send e-post til Kartverket" href="mailto:post@kartverket.no">
+                  <Link
+                    fontWeight="bold"
+                    aria-label="Send e-post til Kartverket"
+                    href={"mailto:" + contactInfoEmailAddress}
+                  >
                     {" "}
-                    post@kartverket.no{" "}
+                    {contactInfoEmailAddress}{" "}
                   </Link>
                 </Text>
                 <Link