+++ packages/react/src/index.ts
@@ -58,3 +58,5 @@ export * from "./datepicker";
 export * from "./file-upload";
 export * from "./logo";
 export * from "./search-async";
+export * from "./footer";
+export * from "./header";