+++ packages/react/src/theme/components/button.ts
@@ -271,13 +271,13 @@ export const buttonTheme = defineStyleConfig({
     },
   },
   variants: {
-    solid: variantSolid,
-    outline: variantOutline,
-    link: variantLink,
+    primary: variantSolid,
+    secondary: variantOutline,
+    tertiary: variantLink,
     ghost: variantGhost,
   },
   defaultProps: {
-    variant: "solid",
+    variant: "primary",
     size: "md",
   },
 });